publisher/requirements.txt
confluent-kafka==2.6.1
watchdog==5.0.3
python-dateutil==2.9.0.post0
